refactor searchscreen & wastbasketscreen
