+++ app/src/main/java/org/grupp18/sortsmart/ui/screen/SearchScreen.kt
@@ -1,124 +1,195 @@
-package org.grupp18.sortsmart.ui.screen
+package org.grupp18.sortsmart
 
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Close
 import androidx.compose.material.icons.filled.History
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.unit.dp
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.lifecycle.viewmodel.compose.viewModel
 import coil.ImageLoader
 import coil.compose.AsyncImage
 import coil.decode.SvgDecoder
 import coil.request.ImageRequest
-import org.grupp18.sortsmart.ui.theme.SortSmartGreen
-import org.grupp18.sortsmart.viewmodel.SearchViewModel
+import kotlinx.coroutines.CancellationException
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
+import org.grupp18.sortsmart.data.api.ItemRetrofitClient
+import org.grupp18.sortsmart.data.model.ItemDetail
+import org.grupp18.sortsmart.data.model.SearchItem
 
 @Composable
 fun SearchScreen(
     modifier: Modifier = Modifier,
     onClose: () -> Unit,
-    viewModel: SearchViewModel = viewModel()
+    onAddToBasket: (ItemDetail) -> Unit
 ) {
-    val query by viewModel.query.collectAsStateWithLifecycle()
-    val suggestions by viewModel.suggestions.collectAsStateWithLifecycle()
-    val selectedItem by viewModel.selectedItem.collectAsStateWithLifecycle()
-    val error by viewModel.error.collectAsStateWithLifecycle()
+    var searchText by remember { mutableStateOf("") }
+    var suggestions by remember { mutableStateOf<List<SearchItem>>(emptyList()) }
+    var selectedItem by remember { mutableStateOf<ItemDetail?>(null) }
+    var errorMessage by remember { mutableStateOf<String?>(null) }
+    var showSuggestions by remember { mutableStateOf(false) }
 
     val context = LocalContext.current
+    val scope = rememberCoroutineScope()
+
+    // Image loader with SVG support for category icons
     val imageLoader = remember {
-        ImageLoader.Builder(context).components { add(SvgDecoder.Factory()) }.build()
+        ImageLoader.Builder(context)
+            .components {
+                add(SvgDecoder.Factory())
+            }
+            .build()
     }
 
+    // Loads full item details when the user selects a search result
+    fun selectItem(item: SearchItem) {
+        searchText = item.name
+        suggestions = emptyList()
+        showSuggestions = false
+        selectedItem = null
+
+        scope.launch {
+            try {
+                selectedItem = ItemRetrofitClient.apiService.getItemBySlug(item.slug)
+                errorMessage = null
+            } catch (e: Exception) {
+                errorMessage = e.localizedMessage ?: "Could not load item details"
+            }
+        }
+    }
+
+    // Updates search suggestions when the search text changes
+    LaunchedEffect(searchText) {
+        if (
+            searchText.isBlank() ||
+            searchText.length < 2 ||
+            selectedItem?.name == searchText
+        ) {
+            suggestions = emptyList()
+            errorMessage = null
+            return@LaunchedEffect
+        }
+
+        delay(400)
+
+        try {
+            val response = ItemRetrofitClient.apiService.searchItems(searchText)
+            suggestions = response.results
+            showSuggestions = true
+            errorMessage = null
+        } catch (e: CancellationException) {
+            throw e
+        } catch (e: Exception) {
+            e.printStackTrace()
+            errorMessage = e.localizedMessage ?: "Unknown error"
+            suggestions = emptyList()
+        }
+    }
+
+    // Main layout for search input, suggestions, errors, and item details
     Column(
         modifier = modifier
             .fillMaxSize()
             .padding(24.dp)
     ) {
-
-        Row(modifier = Modifier.fillMaxWidth(), verticalAlignment = Alignment.CenterVertically) {
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
             OutlinedTextField(
-                value = query,
-                onValueChange = { viewModel.onQueryChange(it) },
+                value = searchText,
+                onValueChange = {
+                    searchText = it
+                    selectedItem = null
+                    showSuggestions = true
+                },
                 placeholder = { Text("What do you want to recycle?") },
                 modifier = Modifier.weight(1f),
                 singleLine = true,
-                keyboardOptions = KeyboardOptions(imeAction = ImeAction.Search),
+                keyboardOptions = KeyboardOptions(
+                    imeAction = ImeAction.Search
+                ),
                 keyboardActions = KeyboardActions(
-                    onSearch = { suggestions.firstOrNull()?.let { viewModel.selectItem(it) } }
+                    onSearch = {
+                        if (suggestions.isNotEmpty()) {
+                            selectItem(suggestions.first())
+                        }
+                    }
                 )
             )
-            IconButton(onClick = {
-                viewModel.clearSelection()
-                onClose()
-            }) {
-                Icon(Icons.Default.Close, contentDescription = "Close")
+
+            IconButton(onClick = onClose) {
+                Icon(
+                    imageVector = Icons.Default.Close,
+                    contentDescription = "Close"
+                )
             }
         }
 
-        Spacer(Modifier.height(24.dp))
+        Spacer(modifier = Modifier.height(24.dp))
 
-        error?.let {
-            Text("Error: $it", color = MaterialTheme.colorScheme.error)
-            Spacer(Modifier.height(12.dp))
+        errorMessage?.let {
+            Text(
+                text = "Error: $it",
+                color = MaterialTheme.colorScheme.error
+            )
+            Spacer(modifier = Modifier.height(12.dp))
         }
 
-        if (suggestions.isNotEmpty() && selectedItem == null) {
-            Text("Suggested", style = MaterialTheme.typography.titleMedium)
-            Spacer(Modifier.height(12.dp))
+        if (showSuggestions && suggestions.isNotEmpty() && selectedItem == null) {
+            Text(
+                text = "Suggested",
+                style = MaterialTheme.typography.titleMedium
+            )
+
+            Spacer(modifier = Modifier.height(12.dp))
+
             suggestions.forEach { item ->
                 Row(
                     modifier = Modifier
                         .fillMaxWidth()
-                        .clickable { viewModel.selectItem(item) }
+                        .clickable {
+                            selectItem(item)
+                        }
                         .padding(vertical = 12.dp),
                     verticalAlignment = Alignment.CenterVertically
                 ) {
-                    Icon(Icons.Default.History, contentDescription = null)
-                    Spacer(Modifier.width(12.dp))
+                    Icon(
+                        imageVector = Icons.Default.History,
+                        contentDescription = null
+                    )
+
+                    Spacer(modifier = Modifier.width(12.dp))
+
                     Text(item.name)
                 }
             }
         }
 
         selectedItem?.let { item ->
-            Spacer(Modifier.height(24.dp))
+            Spacer(modifier = Modifier.height(24.dp))
+
             Card(
                 modifier = Modifier.fillMaxWidth(),
                 shape = RoundedCornerShape(20.dp),
                 elevation = CardDefaults.cardElevation(4.dp)
             ) {
-                Column(Modifier.padding(16.dp)) {
-                    Row(verticalAlignment = Alignment.CenterVertically) {
+                Column(
+                    modifier = Modifier.padding(16.dp)
+                ) {
+                    Row(
+                        verticalAlignment = Alignment.CenterVertically
+                    ) {
                         Box(
                             modifier = Modifier
                                 .size(56.dp)
@@ -135,29 +206,40 @@ fun SearchScreen(
                                 modifier = Modifier.size(48.dp)
                             )
                         }
-                        Spacer(Modifier.width(12.dp))
-                        Text(item.name, style = MaterialTheme.typography.titleLarge)
+
+                        Spacer(modifier = Modifier.width(12.dp))
+
+                        Text(
+                            text = item.name,
+                            style = MaterialTheme.typography.titleLarge
+                        )
                     }
 
-                    Spacer(Modifier.height(16.dp))
+                    Spacer(modifier = Modifier.height(16.dp))
+
                     Text("Category", style = MaterialTheme.typography.titleSmall)
                     Text(item.category?.name ?: "Unknown")
 
-                    Spacer(Modifier.height(12.dp))
+                    Spacer(modifier = Modifier.height(12.dp))
+
                     Text("Leave at", style = MaterialTheme.typography.titleSmall)
                     Text(item.leaveAt ?: "Unknown")
 
-                    Spacer(Modifier.height(12.dp))
+                    Spacer(modifier = Modifier.height(12.dp))
+
                     Text("Processing", style = MaterialTheme.typography.titleSmall)
                     Text(item.processing ?: "No information available")
 
-                    Spacer(Modifier.height(20.dp))
+                    Spacer(modifier = Modifier.height(20.dp))
+
                     Button(
-                        onClick = { /* TODO: add to waste basket */ },
+                        onClick = {
+                            onAddToBasket(item)
+                        },
                         modifier = Modifier.fillMaxWidth(),
                         colors = ButtonDefaults.buttonColors(
-                            containerColor = SortSmartGreen,
-                            contentColor = androidx.compose.ui.graphics.Color.White
+                            containerColor = Color(0xFF2E7D32),
+                            contentColor = Color.White
                         ),
                         shape = RoundedCornerShape(16.dp)
                     ) {
